+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/TigerProxy.java
@@ -73,8 +73,8 @@ public class TigerProxy extends AbstractTigerProxy implements AutoCloseable {
     public static final String CA_CERT_ALIAS = "caCert";
     private final List<DynamicTigerKeyAndCertificateFactory> tlsFactories = new ArrayList<>();
     private final List<Consumer<Throwable>> exceptionListeners = new ArrayList<>();
-    private final MockServer mockServer;
-    private final MockServerClient mockServerClient;
+    private MockServer mockServer;
+    private MockServerClient mockServerClient;
     @Getter
     private final MockServerToRbelConverter mockServerToRbelConverter;
     private final Map<String, TigerRoute> tigerRouteMap = new HashMap<>();
@@ -91,35 +91,9 @@ public class TigerProxy extends AbstractTigerProxy implements AutoCloseable {
     public TigerProxy(final TigerProxyConfiguration configuration) {
         super(configuration);
 
-        KeyAndCertificateFactoryFactory.setCustomKeyAndCertificateFactorySupplier(buildKeyAndCertificateFactory());
-
         mockServerToRbelConverter = new MockServerToRbelConverter(getRbelLogger().getRbelConverter());
-        Configuration mockServerConfiguration = Configuration.configuration();
-        mockServerConfiguration.forwardProxyTLSX509CertificatesTrustManagerType(
-            ForwardProxyTLSX509CertificatesTrustManager.ANY);
-        mockServerConfiguration.maxLogEntries(0);
-        if (StringUtils.isNotEmpty(configuration.getProxyLogLevel())) {
-            mockServerConfiguration.logLevel(configuration.getProxyLogLevel());
-        } else {
-            mockServerConfiguration.logLevel("WARN");
-        }
 
-        customizeSslSuitesIfApplicable();
-
-        final Optional<ProxyConfiguration> forwardProxyConfig = ProxyConfigurationConverter.convertForwardProxyConfigurationToMockServerConfiguration(configuration);
-        outputForwardProxyConfigLogs(forwardProxyConfig);
-
-        if (configuration.getDirectReverseProxy() == null) {
-            mockServer = forwardProxyConfig
-                .map(proxyConfiguration -> new MockServer(proxyConfiguration, configuration.getPortAsArray()))
-                .orElseGet(() -> new MockServer(configuration.getPortAsArray()));
-        } else {
-            mockServer = spawnDirectInverseTigerProxy(mockServerConfiguration, forwardProxyConfig);
-        }
-        log.info("Proxy started on port {}", mockServer.getLocalPort());
-
-        mockServerClient = new MockServerClient("localhost", mockServer.getLocalPort());
-        addRoutesToTigerProxy();
+        bootMockServer();
 
         if (!configuration.isSkipTrafficEndpointsSubscription()) {
             subscribeToTrafficEndpoints(configuration);
@@ -134,12 +108,78 @@ public class TigerProxy extends AbstractTigerProxy implements AutoCloseable {
                 getRbelLogger().getRbelModifier().addModification(modification);
             }
         }
+    }
 
-        if (configuration.isActivateForwardAllLogging()) {
+    /**
+     * Will restart the internal mockserver. This can be done to force a reload of the configured TLS-identities.
+     * This will block until the mockserver is running again. Background connections will most likely be interrupted.
+     * Routes will be re-added to the server, but the ID's will change.
+     * 
+     */
+    public void restartMockserver() {
+        if (getTigerProxyConfiguration().getProxyPort() == null) {
+            getTigerProxyConfiguration().setProxyPort(mockServer.getLocalPort());
+        }
+        mockServer.stop();
+        mockServerClient.stop();
+        var originalRoutes = Collections.unmodifiableMap(tigerRouteMap);
+        tigerRouteMap.clear();
+        bootMockServer();
+        originalRoutes.values().stream()
+            .filter(r -> !r.isInternalRoute())
+            .forEach(r -> {
+                try {
+                    addRoute(r);
+                } catch (RuntimeException e) {
+                    //swallow
+                    log.trace("Ignored exception during re-adding of routes", e);
+                }
+            }) ;
+    }
+
+    private void bootMockServer() {
+        createNewMockServer();
+        buildMockServerClient();
+    }
+
+    private void buildMockServerClient() {
+        mockServerClient = new MockServerClient("localhost", mockServer.getLocalPort());
+
+        if (getTigerProxyConfiguration().isActivateForwardAllLogging()) {
             mockServerClient.when(request()
                     .withPath(".*"), Times.unlimited(), TimeToLive.unlimited(), Integer.MIN_VALUE)
                 .forward(new ForwardAllCallback(this));
         }
+        addRoutesToTigerProxy();
+    }
+
+    private void createNewMockServer() {
+        KeyAndCertificateFactoryFactory.setCustomKeyAndCertificateFactorySupplier(buildKeyAndCertificateFactory());
+
+        Configuration mockServerConfiguration = Configuration.configuration();
+        mockServerConfiguration.forwardProxyTLSX509CertificatesTrustManagerType(
+            ForwardProxyTLSX509CertificatesTrustManager.ANY);
+        mockServerConfiguration.maxLogEntries(0);
+        if (StringUtils.isNotEmpty(getTigerProxyConfiguration().getProxyLogLevel())) {
+            mockServerConfiguration.logLevel(getTigerProxyConfiguration().getProxyLogLevel());
+        } else {
+            mockServerConfiguration.logLevel("WARN");
+        }
+
+        customizeSslSuitesIfApplicable();
+
+        final Optional<ProxyConfiguration> forwardProxyConfig = ProxyConfigurationConverter.convertForwardProxyConfigurationToMockServerConfiguration(
+            getTigerProxyConfiguration());
+        outputForwardProxyConfigLogs(forwardProxyConfig);
+
+        if (getTigerProxyConfiguration().getDirectReverseProxy() == null) {
+            mockServer = forwardProxyConfig
+                .map(proxyConfiguration -> new MockServer(proxyConfiguration, getTigerProxyConfiguration().getPortAsArray()))
+                .orElseGet(() -> new MockServer(getTigerProxyConfiguration().getPortAsArray()));
+        } else {
+            mockServer = spawnDirectInverseTigerProxy(mockServerConfiguration, forwardProxyConfig);
+        }
+        log.info("Proxy '{}' started on port {}", getName().orElse("?"), mockServer.getLocalPort());
     }
 
     private void addRoutesToTigerProxy() {